requirements.txt
drf-yasg==1.21.7
channels==4.1.0
djangorestframework==3.14.0
Pillow==9.5.0
python-dotenv=1.0.1
dj_database_url==
django-cors-headers==4.11.0
daphne==2.21
gunicorn==22.0.0
